Add default release-drafter template
https://github.com/release-drafter/release-drafter (available through
sicpa-foundations/actions/external) needs a template to work.
The action uses probot (and calls `.config()` to load the template),
and probot makes it possible to define default configurations in the `.github`
repository of the organization as fallback.

This is better explained here:
https://probot.github.io/docs/best-practices/#store-configuration-in-the-repository

This commit adds this default template, that should be fine for all dlab
repositories. Just add the release-drafter action to a workflow !